✨FIX: BookmarkStar 컴포넌트와 합침
