+++ src/components/main/Diary/Diary.jsx
@@ -1,13 +1,60 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { darken } from "polished";
+import { useDispatch } from "react-redux";
 import styled, { css } from "styled-components";
-import BookmarkStar from "./BookmarkStar";
+import { mainApi } from "../../../apis/axios";
+import { ErrorModal } from "../../../redux/modules/UISlice";
+import { IoMdBookmark } from "react-icons/io";
+import { AiOutlineStar } from "react-icons/ai";
+import { AiFillStar } from "react-icons/ai";
 
 const Diary = (props) => {
   const { size, bgColor, onClick, bookmark, diaryId } = props;
 
+  const queryClient = useQueryClient();
+  const dispatch = useDispatch();
+
+  const { mutate } = useMutation(["diary"], (diaryId) => mainApi.bookmark(diaryId), {
+    onError: (error) => {
+      const status = error?.response.request.status;
+      if (status === 401) dispatch(ErrorModal({ isModal: true, bigTxt: "권한이 없습니다." }));
+      else if (status === 404)
+        dispatch(
+          ErrorModal({
+            isModal: true,
+            bigTxt: "존재하지 않는 다이어리입니다.",
+          })
+        );
+      else if (status === 500)
+        dispatch(
+          ErrorModal({
+            isModal: true,
+            bigTxt: "북마크 저장 및 삭제에 실패했습니다.",
+          })
+        );
+    },
+    onSuccess: () => {
+      const diaryData = queryClient.getQueryData(["main"])?.diaries;
+      queryClient.setQueryData(["main"], {
+        diaries: diaryData?.map((diary) =>
+          diary.diaryId === diaryId ? { ...diary, bookmark: diary.bookmark === 0 ? 1 : 0 } : diary
+        ),
+      });
+    },
+  });
+
+  const bookmarkHandler = (event) => {
+    event.stopPropagation();
+    mutate(diaryId);
+  };
+
   return (
     <DiaryIcon size={size} bgcolor={bgColor ? bgColor : "#E9E9E9"} onClick={onClick}>
-      <BookmarkStar bookmark={bookmark} diaryId={diaryId} />
+      <BookmarkStarDiv onClick={bookmarkHandler}>
+        {bookmark === 0 ? <AiOutlineStar /> : <AiFillStar className="star" />}
+        {/* {bookmark === 0 ? <AiOutlineStar /> : <AiFillStar className="star" />}
+        <IoMdBookmark size={size} className={bookmark === 1 && "bookmarked"} /> */}
+      </BookmarkStarDiv>
       <div className="diaryHolder" />
     </DiaryIcon>
   );
@@ -86,4 +133,24 @@ const DiaryIcon = styled.div`
       }
     }}
   }
+`;
+
+// const BookmarkStarDiv = styled.div`
+//   float: left;
+//   margin: -13% 0 0 5%;
+//   color: white;
+//   font-size: 6rem;
+//   cursor: pointer;
+//   .bookmarked {
+//     color: #fffa84;
+//   }
+// `;
+
+const BookmarkStarDiv = styled.div`
+  float: left;
+  margin: 10% 0 0 10%;
+  width: 3rem;
+  font-size: 2.5rem;
+  color: #fdcb6e;
+  cursor: pointer;
 `;
